refactor: separate dev/prd handler/middleware
