+++ examples/08_cookies/start.js
@@ -2,7 +2,7 @@ import url from 'node:url';
 import path from 'node:path';
 import express from 'express';
 import cookieParser from 'cookie-parser';
-import { connectMiddleware } from 'waku';
+import { connectPrdMiddleware } from 'waku';
 
 const withSsr = process.argv[2] === '--with-ssr';
 
@@ -16,9 +16,8 @@ const root = path.join(
 const app = express();
 app.use(cookieParser());
 app.use(
-  connectMiddleware({
+  connectPrdMiddleware({
     config,
-    command: 'start',
     unstable_prehook: (req) => {
       return { count: Number(req.orig.cookies.count) || 0 };
     },
