+++ packages/waku/package.json
@@ -64,14 +64,16 @@
   "engines": {
     "node": ">=18.0.0"
   },
+  "dependencies": {
+    "@swc/core": "1.3.96",
+    "@vitejs/plugin-react": "4.2.0",
+    "vite": "5.0.4"
+  },
   "devDependencies": {
     "@hono/node-server": "^1.2.2",
     "@swc/cli": "^0.1.62",
-    "@swc/core": "1.3.96",
-    "@vitejs/plugin-react": "4.2.0",
     "hono": "^3.10.0",
-    "rollup": "^4.3.0",
-    "vite": "5.0.4"
+    "rollup": "^4.3.0"
   },
   "peerDependencies": {
     "@hono/node-server": "^1.2.2",
@@ -86,16 +88,7 @@
     "@hono/node-server": {
       "optional": true
     },
-    "@swc/core": {
-      "optional": true
-    },
-    "@vitejs/plugin-react": {
-      "optional": true
-    },
     "hono": {
-      "optional": true
-    },
-    "vite": {
       "optional": true
     }
   }
