+++ packages/waku/src/main.ts
@@ -1,9 +1,11 @@
-export { honoMiddleware } from './lib/middleware/hono.js';
-export { connectMiddleware } from './lib/middleware/connect.js';
-export { createHandler as unstable_createHandler } from './lib/rsc/handler.js';
+export { honoDevMiddleware, honoPrdMiddleware } from './lib/middleware/hono.js';
 
-import type { build as buildOrig } from './lib/builder.js';
+export {
+  connectDevMiddleware,
+  connectPrdMiddleware,
+} from './lib/middleware/connect.js';
 
-export async function build(...args: Parameters<typeof buildOrig>) {
-  return (await import('./lib/builder.js')).build(...args);
-}
+export { createDevHandler as unstable_createDevHandler } from './lib/rsc/dev-handler.js';
+export { createPrdHandler as unstable_createPrdHandler } from './lib/rsc/prd-handler.js';
+
+export { build } from './lib/builder.js';
